Finished Commands

+++ RoboRIO/src/main/java/org/usfirst/frc/team449/robot/subsystem/interfaces/Position/SubsystemPosition.java
@@ -3,10 +3,10 @@ package org.usfirst.frc.team449.robot.subsystem.interfaces.Position;
 public interface SubsystemPosition {
 
 	// Method for setting the position to a value.
-	public void setPosition(int value);
+	public void setPosition(double value);
 
 	// Method for setting the motor output to a value.
-	public void setMotorOutput(int value);
+	public void setMotorOutput(double value);
 
 	// Method for returning the state of the reverse limit switch.
 	public boolean getReverseLimit();
